requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.8
django-allauth==0.35.0
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
oauthlib==2.0.6
Pygments==2.2.0
PyMySQL==0.8.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
uritemplate==3.0.0
urllib3==1.22
